# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Absher Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visas:
    get:
      operationId: listVisas
      tags: [visas]
      summary: List visa information by country
      responses:
        "200":
          description: List of visa entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visa"
    post:
      operationId: createVisa
      tags: [visas]
      summary: Create visa entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visa"
  /visas/{id}:
    get:
      operationId: getVisa
      tags: [visas]
      summary: Get visa details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visa details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visa"
    patch:
      operationId: updateVisa
      tags: [visas]
      summary: Update visa entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visa"
    delete:
      operationId: deleteVisa
      tags: [visas]
      summary: Delete visa entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /visa-applications:
    get:
      operationId: listVisaApplications
      tags: [visa-applications]
      summary: List visa applications (admin)
      parameters:
        - name: visaId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [received, under_review, awaiting_documents, documents_uploaded, sent_to_embassy, processing, issued, completed, rejected]
      responses:
        "200":
          description: List of visa applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisaApplication"
    post:
      operationId: createVisaApplication
      tags: [visa-applications]
      summary: Submit a new visa application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisaApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisaApplication"
        "422":
          description: Applicant is not eligible for this visa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /visa-applications/{id}:
    get:
      operationId: getVisaApplication
      tags: [visa-applications]
      summary: Get a visa application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visa application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisaApplication"
    patch:
      operationId: updateVisaApplication
      tags: [visa-applications]
      summary: Update a visa application status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisaApplicationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisaApplication"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Visa:
      type: object
      required: [id, countryAr, countryEn, visaType, processingDays, fee, currency, entryType, status, isActive, createdAt]
      properties:
        id: { type: integer }
        countryAr: { type: string }
        countryEn: { type: string }
        countryCode: { type: string }
        visaType: { type: string }
        category:
          type: string
          enum: [tourist, business, medical, visit, study, umrah]
        descriptionAr: { type: ["string", "null"] }
        descriptionEn: { type: ["string", "null"] }
        requirements: { type: string }
        documents: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        processingDays: { type: integer }
        fee: { type: number }
        currency: { type: string }
        stayDuration: { type: ["integer", "null"] }
        validityDays: { type: ["integer", "null"] }
        entryType:
          type: string
          enum: [single, multiple, transit]
        entryCount: { type: ["integer", "null"] }
        allowedNationalities:
          type: array
          items: { type: string }
        blockedNationalities:
          type: array
          items: { type: string }
        imageUrl: { type: ["string", "null"] }
        status:
          type: string
          enum: [available, suspended, closed]
        isActive: { type: boolean }
        acceptsGccResidency: { type: boolean }
        acceptsSchengenResidency: { type: boolean }
        acceptsUkResidency: { type: boolean }
        acceptsUsVisa: { type: boolean }
        acceptsCanadaResidency: { type: boolean }
        acceptsAustraliaResidency: { type: boolean }
        requiresPassportImage: { type: boolean }
        requiresPersonalPhoto: { type: boolean }
        requiresResidencyImage: { type: boolean }
        requiresVisaImage: { type: boolean }
        ineligibleMessageAr: { type: ["string", "null"] }
        ineligibleMessageEn: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    VisaInput:
      type: object
      required: [countryAr, countryEn, visaType, processingDays, fee, currency]
      properties:
        countryAr: { type: string }
        countryEn: { type: string }
        countryCode: { type: string }
        visaType: { type: string }
        category:
          type: string
          enum: [tourist, business, medical, visit, study, umrah]
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        requirements: { type: string }
        documents: { type: string }
        notes: { type: string }
        processingDays: { type: integer }
        fee: { type: number }
        currency: { type: string }
        stayDuration: { type: integer }
        validityDays: { type: integer }
        entryType:
          type: string
          enum: [single, multiple, transit]
        entryCount: { type: integer }
        allowedNationalities:
          type: array
          items: { type: string }
        blockedNationalities:
          type: array
          items: { type: string }
        imageUrl: { type: string }
        status:
          type: string
          enum: [available, suspended, closed]
        isActive: { type: boolean }
        acceptsGccResidency: { type: boolean }
        acceptsSchengenResidency: { type: boolean }
        acceptsUkResidency: { type: boolean }
        acceptsUsVisa: { type: boolean }
        acceptsCanadaResidency: { type: boolean }
        acceptsAustraliaResidency: { type: boolean }
        requiresPassportImage: { type: boolean }
        requiresPersonalPhoto: { type: boolean }
        requiresResidencyImage: { type: boolean }
        requiresVisaImage: { type: boolean }
        ineligibleMessageAr: { type: string }
        ineligibleMessageEn: { type: string }
    VisaUpdate:
      type: object
      properties:
        countryAr: { type: string }
        countryEn: { type: string }
        countryCode: { type: string }
        visaType: { type: string }
        category:
          type: string
          enum: [tourist, business, medical, visit, study, umrah]
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        requirements: { type: string }
        documents: { type: string }
        notes: { type: string }
        processingDays: { type: integer }
        fee: { type: number }
        currency: { type: string }
        stayDuration: { type: integer }
        validityDays: { type: integer }
        entryType:
          type: string
          enum: [single, multiple, transit]
        entryCount: { type: integer }
        allowedNationalities:
          type: array
          items: { type: string }
        blockedNationalities:
          type: array
          items: { type: string }
        imageUrl: { type: string }
        status:
          type: string
          enum: [available, suspended, closed]
        isActive: { type: boolean }
        acceptsGccResidency: { type: boolean }
        acceptsSchengenResidency: { type: boolean }
        acceptsUkResidency: { type: boolean }
        acceptsUsVisa: { type: boolean }
        acceptsCanadaResidency: { type: boolean }
        acceptsAustraliaResidency: { type: boolean }
        requiresPassportImage: { type: boolean }
        requiresPersonalPhoto: { type: boolean }
        requiresResidencyImage: { type: boolean }
        requiresVisaImage: { type: boolean }
        ineligibleMessageAr: { type: string }
        ineligibleMessageEn: { type: string }
    VisaApplication:
      type: object
      required: [id, visaId, eligibilityPath, fullName, nationality, passportNumber, passportIssueDate, passportExpiryDate, dateOfBirth, gender, email, phone, agreedToTerms, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        visaId: { type: integer }
        userId: { type: ["string", "null"] }
        eligibilityPath:
          type: string
          enum: [gcc, alternative, direct]
        gccCountry: { type: ["string", "null"] }
        alternativeRegion: { type: ["string", "null"] }
        fullName: { type: string }
        nationality: { type: string }
        passportNumber: { type: string }
        passportIssueDate: { type: string }
        passportExpiryDate: { type: string }
        dateOfBirth: { type: string }
        gender:
          type: string
          enum: [male, female]
        email: { type: string }
        phone: { type: string }
        passportImageUrl: { type: ["string", "null"] }
        personalPhotoUrl: { type: ["string", "null"] }
        residencyImageUrl: { type: ["string", "null"] }
        visaImageUrl: { type: ["string", "null"] }
        agreedToTerms: { type: boolean }
        status:
          type: string
          enum: [received, under_review, awaiting_documents, documents_uploaded, sent_to_embassy, processing, issued, completed, rejected]
        adminNotes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    VisaApplicationInput:
      type: object
      required: [visaId, eligibilityPath, fullName, nationality, passportNumber, passportIssueDate, passportExpiryDate, dateOfBirth, gender, email, phone, agreedToTerms]
      properties:
        visaId: { type: integer }
        eligibilityPath:
          type: string
          enum: [gcc, alternative, direct]
        gccCountry: { type: string }
        alternativeRegion: { type: string }
        fullName: { type: string }
        nationality: { type: string }
        passportNumber: { type: string }
        passportIssueDate: { type: string }
        passportExpiryDate: { type: string }
        dateOfBirth: { type: string }
        gender:
          type: string
          enum: [male, female]
        email: { type: string }
        phone: { type: string }
        passportImageUrl: { type: string }
        personalPhotoUrl: { type: string }
        residencyImageUrl: { type: string }
        visaImageUrl: { type: string }
        agreedToTerms: { type: boolean }
    VisaApplicationStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [received, under_review, awaiting_documents, documents_uploaded, sent_to_embassy, processing, issued, completed, rejected]
        adminNotes: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT